Changes to sys_toolkit.tmpdir functionality
Make the default implementation to just skip handling of ramdisk
volume to the temporary directory by default, insted of requiring
subclasses to do this: seems only darwin does this nicely right now.

Also add method to list files inside the temporary directory, skipping
symlinks to files outside of the directory. This is to be used when
shredding directory contents later.